feat: abstracted PriceLabel
