+++ src/shared/components/PriceLabel/priceLabel.scss
@@ -1,0 +1,38 @@
+@import '../../../app/globals.scss';
+
+.customLabel {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+
+  width: 32%;
+  padding: 2px 0;
+
+  text-align: center;
+  font-size: var(--xxs);
+  color: var(--c-light-violet);
+  background-color: var(--c-ultra-violet);
+
+  &:hover {
+    cursor: pointer;
+  }
+
+  &:has(input:checked) {
+    background-color: var(--c-african-violet);
+    color: var(--c-linen-texts);
+  }
+
+  @include for-tablet-portrait-up {
+    font-size: 13px;
+  }
+
+  input {
+    position: absolute;
+    width: 0;
+    opacity: 0;
+    margin: 0;
+    pointer-events: none;
+  }
+}
